refactor(destination-redshift-v2): reorganize RedshiftSpecification into config package
- Move RedshiftSpecification from spec to config package for better organization
- Remove unused import in RedshiftBeanFactory
- Simplify RedshiftConfiguration by removing helper properties and comments
- Ensure clean build with all refactored imports
