+++ airbyte-integrations/connectors/destination-redshift-v2/src/main/kotlin/io/airbyte/integrations/destination/redshift2/config/RedshiftConfiguration.kt
@@ -6,13 +6,10 @@ package io.airbyte.integrations.destination.redshift2.config
 
 import io.airbyte.cdk.load.command.DestinationConfiguration
 import io.airbyte.cdk.load.command.DestinationConfigurationFactory
-import io.airbyte.integrations.destination.redshift2.spec.RedshiftSpecification
-import io.airbyte.integrations.destination.redshift2.spec.S3StagingConfig
 import jakarta.inject.Singleton
 
 /**
  * Typed configuration for Redshift destination.
- * Provides computed properties for JDBC URL and S3 staging configuration.
  */
 data class RedshiftConfiguration(
     val host: String,
@@ -25,39 +22,11 @@ data class RedshiftConfiguration(
     val uploadingMethod: S3StagingConfig?
 ) : DestinationConfiguration() {
     
-    /** Computed JDBC URL for Redshift connection. */
     val jdbcUrl: String
         get() {
             val baseUrl = "jdbc:redshift://$host:$port/$database"
-            return if (jdbcUrlParams.isNullOrBlank()) {
-                baseUrl
-            } else {
-                "$baseUrl?$jdbcUrlParams"
-            }
+            return if (jdbcUrlParams.isNullOrBlank()) baseUrl else "$baseUrl?$jdbcUrlParams"
         }
-    
-    /** Whether S3 staging is configured. */
-    val hasS3Staging: Boolean
-        get() = uploadingMethod != null
-    
-    /** S3 staging configuration accessors. */
-    val s3BucketName: String?
-        get() = uploadingMethod?.s3BucketName
-    
-    val s3BucketPath: String?
-        get() = uploadingMethod?.s3BucketPath
-    
-    val s3BucketRegion: String?
-        get() = uploadingMethod?.s3BucketRegion
-    
-    val s3AccessKeyId: String?
-        get() = uploadingMethod?.accessKeyId
-    
-    val s3SecretAccessKey: String?
-        get() = uploadingMethod?.secretAccessKey
-    
-    val purgeStagingData: Boolean
-        get() = uploadingMethod?.purgeStagingData ?: true
 }
 
 /**
